Added fatal error message box for the.. fatal stuff. Windows isn't really good at conveying error messages, converted actions into a C-style array because I've had a ton of problems with the vector approach (it gets zeroed when it shouldn't be..)

+++ src/string.cpp
@@ -177,7 +177,7 @@ String String::substr (long a, long b) const {
 int String::first (const char* c, int a) const {
 	unsigned int r = 0;
 	unsigned int index = 0;
-	for (; a < len (); a++) {
+	for (; a < (int) len (); a++) {
 		if (m_string[a] == c[r]) {
 			if (r == 0)
 				index = a;
